+++ Source/Public/Window.h
@@ -10,20 +10,20 @@ class Window
 {
 public:
 
-	static void Create(int width, int height, const std::string& title);
+	static void Create(float width, float height, const std::string& title);
 	static void Destroy();
 	static void Render();
 	static bool CloseRequested();
 
-	static int GetWidth() { return width_; };
-	static int GetHeight() { return height_; };
+	static float GetWidth() { return width_; };
+	static float GetHeight() { return height_; };
 	static const std::string& GetTitle() { return title_; };
 	static void SetFullScreen(bool value);
 
 private:
 
-	static int width_;
-	static int height_;
+	static float width_;
+	static float height_;
 	static std::string title_;
 };
